+++ airflow_ml_dags/images/airflow-train/train.py
@@ -1,14 +1,9 @@
-import pathlib
-import pandas as pd
 import click
 import os
 
-from sklearn.pipeline import Pipeline
-from sklearn.preprocessing import StandardScaler
-from sklearn.linear_model import LogisticRegression
-
 import mlflow
 
+from log_reg import train
 
 @click.command()
 @ click.option("--train_dir")
@@ -28,21 +23,9 @@ def mlflow_train(train_dir: str, exp_name: str):
 
     with mlflow.start_run() as run:
         pipeline = train(train_dir)
-
-
-def train(train_dir: str):
-    train_datapath = pathlib.Path(train_dir) / "data.csv"
-    target_datapath = train_datapath.parent / "target.csv"
-
-    features = pd.read_csv(train_datapath, encoding="utf-8")
-    target = pd.read_csv(target_datapath, encoding="utf-8")
-
-    pipeline = Pipeline([("scaler", StandardScaler()),
-                         ("predict", LogisticRegression())])
-
-    pipeline.fit(features.to_numpy(), target.to_numpy().ravel())
-
-    return pipeline
+        mlflow.sklearn.log_model(sk_model=pipeline,
+                                 artifact_path='',
+                                 registered_model_name="log_reg")
 
 
 if __name__ == '__main__':
